+++ src/main/java/searching_program/search_product/dto/PromotionDto.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class PromotionDto {
     private Long id;
     private Long itemId;
-    private BigDecimal discountRate;
+    private Long discountRate;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
 }
